Align RSA interrupt enable register/field name for S2/S3 with other chips
