+++ esp32s2/svd/patches/esp32s2.yaml
@@ -342,3 +342,13 @@ SYSTIMER:
 
 USB0:
   _include: ../../../common_patches/usb0.yaml
+
+RSA:
+  _modify:
+    INTERRUPT_ENA:
+      name: INT_ENA
+
+  INT_ENA:
+    _modify:
+      INTERRUPT_ENA:
+        name: INT_ENA
