+++ esp32s3/svd/patches/esp32s3.yaml
@@ -66,21 +66,21 @@ _add:
       usage: registers
     registers:
       GEN_CTRL:
-          description: "FE General Control Register"
-          addressOffset: 0x0090
-          size: 0x2
-          access: read-write
-          fields:
-            IQ_EST_FORCE_PU:
-              description: "Force Power Up for IQ Estimation"
-              bitOffset: 5
-              bitWidth: 1
-              access: read-write
-            IQ_EST_FORCE_PD:
-              description: "Force Power Down for IQ Estimation"
-              bitOffset: 4
-              bitWidth: 1
-              access: read-write
+        description: "FE General Control Register"
+        addressOffset: 0x0090
+        size: 0x2
+        access: read-write
+        fields:
+          IQ_EST_FORCE_PU:
+            description: "Force Power Up for IQ Estimation"
+            bitOffset: 5
+            bitWidth: 1
+            access: read-write
+          IQ_EST_FORCE_PD:
+            description: "Force Power Down for IQ Estimation"
+            bitOffset: 4
+            bitWidth: 1
+            access: read-write
   FE2:
     description: need des
     baseAddress: 0x60005000
@@ -91,21 +91,21 @@ _add:
       usage: registers
     registers:
       TX_INTERP_CTRL:
-          description: "FE2 TX Interpolation Control Register"
-          addressOffset: 0x00f0
-          size: 0x20
-          access: read-write
-          fields:
-            TX_INF_FORCE_PU:
-              description: "Force Power Up field"
-              bitOffset: 10
-              bitWidth: 1
-              access: read-write
-            TX_INF_FORCE_PD:
-              description: "Force Power Down field"
-              bitOffset: 9
-              bitWidth: 1
-              access: read-write
+        description: "FE2 TX Interpolation Control Register"
+        addressOffset: 0x00f0
+        size: 0x20
+        access: read-write
+        fields:
+          TX_INF_FORCE_PU:
+            description: "Force Power Up field"
+            bitOffset: 10
+            bitWidth: 1
+            access: read-write
+          TX_INF_FORCE_PD:
+            description: "Force Power Down field"
+            bitOffset: 9
+            bitWidth: 1
+            access: read-write
 
 _modify:
   SPI?:
@@ -139,7 +139,7 @@ I2C0:
   _modify:
     INT_STATUS:
       name: INT_ST
-  _include: 
+  _include:
     - ../../../common_patches/int_strip.yaml
     - ../../../common_patches/i2c0.yaml
 
@@ -167,7 +167,7 @@ RTC_CNTL:
   _include: ../../../common_patches/int_strip.yaml
 
 LEDC:
-  _include: 
+  _include:
     - ../../../common_patches/ledc_collect.yaml
     - ../../../common_patches/ledc_int.yaml
 
@@ -226,7 +226,7 @@ SPI2:
   _include: ../../../common_patches/spi_dma_int_strip.yaml
 
 SYSTIMER:
-  _include: 
+  _include:
     - ../../../common_patches/int_strip.yaml
     - ../../../common_patches/systimer.yaml
     - ../../../common_patches/systimer_real_target.yaml
@@ -452,3 +452,13 @@ LCD_CAM:
       FallingEdge: [2, Delayed by the falling edge of LCD_CLK]
     _array:
       DOUT*_MODE: {}
+
+RSA:
+  _modify:
+    INTERRUPT_ENA:
+      name: INT_ENA
+
+  INT_ENA:
+    _modify:
+      INTERRUPT_ENA:
+        name: INT_ENA
